+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## master
 * Fix webpack config for demo-app [#9](https://github.com/sham-ui/sham-ui-cli/issues/9)
+* Add eslint support for SFW [#10](https://github.com/sham-ui/sham-ui-cli/issues/10)
+* Remove HMR code from production build
 
 ## 2.0.4
 * Fix `sham-ui-test-helpers` version in app template